# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  description: string;
  topicCount: number;
}

export interface Topic {
  id: number;
  categoryId: number;
  categoryName: string;
  title: string;
  angles: string[];
}

export interface Session {
  id: number;
  sessionKey: string;
  topicId: number;
  topicTitle: string;
  categoryName: string;
  researchSeconds: number;
  speakingSeconds: number;
  createdAt: string;
  /** @nullable */
  recordingUrl: string | null;
  /** @nullable */
  transcript: string | null;
  /** @nullable */
  fillerCount: number | null;
  /** @nullable */
  fillersPerMinute: number | null;
  /** @nullable */
  eyeContactPercent: number | null;
  /** @nullable */
  postureScore: number | null;
  /** @nullable */
  stillnessScore: number | null;
  /** @nullable */
  visualSamples: number | null;
  /** @nullable */
  feedbackSummary: string | null;
  /** @nullable */
  feedbackStrengths: string[] | null;
  /** @nullable */
  feedbackImprovements: string[] | null;
  /** @nullable */
  feedbackNextTip: string | null;
}

export interface SessionInput {
  /** @minLength 1 */
  sessionKey: string;
  topicId: number;
  /** @minimum 0 */
  researchSeconds: number;
  /** @minimum 0 */
  speakingSeconds: number;
  /** @nullable */
  recordingUrl?: string | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  fillerCount?: number | null;
  /** @nullable */
  fillersPerMinute?: number | null;
  /** @nullable */
  eyeContactPercent?: number | null;
  /** @nullable */
  postureScore?: number | null;
  /** @nullable */
  stillnessScore?: number | null;
  /** @nullable */
  visualSamples?: number | null;
  /** @nullable */
  feedbackSummary?: string | null;
  feedbackStrengths?: string[];
  feedbackImprovements?: string[];
  /** @nullable */
  feedbackNextTip?: string | null;
}

export interface SessionSummary {
  sessionCount: number;
  totalSpeakingSeconds: number;
  totalResearchSeconds: number;
  /** @nullable */
  favoriteCategory: string | null;
}

export type ListTopicsParams = {
categoryId?: number;
};

export type GetRandomTopicParams = {
categoryId?: number;
};

export type ListSessionsParams = {
sessionKey: string;
};

export type GetSessionSummaryParams = {
sessionKey: string;
};