# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Session {
  id: number;
  sessionKey: string;
  topicId: number;
  topicTitle: string;
  categoryName: string;
  researchSeconds: number;
  speakingSeconds: number;
  createdAt: string;
  /** @nullable */
  recordingUrl: string | null;
  /** @nullable */
  transcript: string | null;
  /** @nullable */
  fillerCount: number | null;
  /** @nullable */
  fillersPerMinute: number | null;
  /** @nullable */
  eyeContactPercent: number | null;
  /** @nullable */
  postureScore: number | null;
  /** @nullable */
  stillnessScore: number | null;
  /** @nullable */
  visualSamples: number | null;
  /** @nullable */
  feedbackSummary: string | null;
  /** @nullable */
  feedbackStrengths: string[] | null;
  /** @nullable */
  feedbackImprovements: string[] | null;
  /** @nullable */
  feedbackNextTip: string | null;
}