# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  /** @minLength 1 */
  sessionKey: string;
  topicId: number;
  /** @minimum 0 */
  researchSeconds: number;
  /** @minimum 0 */
  speakingSeconds: number;
  /** @nullable */
  recordingUrl?: string | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  fillerCount?: number | null;
  /** @nullable */
  fillersPerMinute?: number | null;
  /** @nullable */
  eyeContactPercent?: number | null;
  /** @nullable */
  postureScore?: number | null;
  /** @nullable */
  stillnessScore?: number | null;
  /** @nullable */
  visualSamples?: number | null;
  /** @nullable */
  feedbackSummary?: string | null;
  feedbackStrengths?: string[];
  feedbackImprovements?: string[];
  /** @nullable */
  feedbackNextTip?: string | null;
}